--- bittorrent-io/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "bittorrent-io"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-io-uring = "0.7"
-libc = "0.2"
-slab = "0.4"
-bytes = "1.8"
-bitvec = { version = "1.0", default-features = false, features = ["alloc"]}
-arbitrary = { version = "1.1.3", features = ["derive"]}
-log = "0.4"
-socket2 = "0.5"
-rand = "0.8"
-lava_torrent = "0.11"
-thiserror = "2"
-sha1 = "0.10.5"
-
-[dev-dependencies]
-env_logger = { workspace = true }
