+++ bittorrent/Cargo.toml
@@ -1,28 +1,21 @@
 [package]
 name = "vortex-bittorrent"
 version = "0.1.0"
-authors = ["Oskar Nehlin"]
 edition = "2021"
-description = "An implementation of the bittorrent protocol built on top of io-uring"
-homepage = "https://github.com/Nehliin/vortex"
-repository = "https://github.com/Nehliin/vortex"
-license = "BSD-3-Clause"
 
 [dependencies]
-bytes = { workspace = true }
-tokio-uring = { workspace = true }
-tokio-util = "0.7.4"
-anyhow = { workspace = true }
-log = { workspace = true }
-lava_torrent = "0.10"
-slotmap = "1"
-parking_lot = { workspace = true }
+io-uring = "0.7"
+libc = "0.2"
+slab = "0.4"
+bytes = "1.8"
 bitvec = { version = "1.0", default-features = false, features = ["alloc"]}
-sha1 = "0.10.5"
-tokio = { workspace = true, features = ["time", "sync", "macros"] }
-rand = { workspace = true }
 arbitrary = { version = "1.1.3", features = ["derive"]}
+log = "0.4"
+socket2 = "0.5"
+rand = "0.8"
+lava_torrent = "0.11"
+thiserror = "2"
+sha1 = "0.10.5"
 
 [dev-dependencies]
 env_logger = { workspace = true }
-rand = { workspace = true }
