+++ bittorrent/src/buf_ring.rs
